refactor(query): standardize query exceptions
