+++ wow-elasticsearch/src/integrationTest/kotlin/me/ahoo/wow/elasticsearch/query/gateway/ElasticsearchSnapshotQueryGatewayTest.kt
@@ -21,7 +21,7 @@ import me.ahoo.wow.api.query.PredicateExpression
 import me.ahoo.wow.api.query.PredicateOperator
 import me.ahoo.wow.api.query.Query
 import me.ahoo.wow.api.query.QueryErrorCode
-import me.ahoo.wow.api.query.QueryException
+import me.ahoo.wow.query.QueryException
 import me.ahoo.wow.api.query.QuerySort
 import me.ahoo.wow.api.query.QuerySortDirection
 import me.ahoo.wow.api.query.SearchExpression
