+++ wow-elasticsearch/src/main/kotlin/me/ahoo/wow/elasticsearch/query/gateway/ElasticsearchQueryMapping.kt
@@ -25,11 +25,11 @@ import me.ahoo.wow.api.query.LogicalField
 import me.ahoo.wow.api.query.PredicateExpression
 import me.ahoo.wow.api.query.PredicateOperator
 import me.ahoo.wow.api.query.QueryErrorCode
-import me.ahoo.wow.api.query.QueryException
 import me.ahoo.wow.api.query.QueryExpression
 import me.ahoo.wow.api.query.QueryStage
 import me.ahoo.wow.api.query.SearchExpression
 import me.ahoo.wow.elasticsearch.IndexNameConverter.toSnapshotIndexName
+import me.ahoo.wow.query.QueryException
 import me.ahoo.wow.query.backend.SecuredQuery
 import me.ahoo.wow.query.schema.QueryFieldSchema
 import me.ahoo.wow.query.schema.QueryValueKind
