+++ wow-mongo/src/main/kotlin/me/ahoo/wow/mongo/query/gateway/MongoSnapshotQueryBackend.kt
@@ -29,7 +29,6 @@ import me.ahoo.wow.api.query.MatchNone
 import me.ahoo.wow.api.query.PredicateExpression
 import me.ahoo.wow.api.query.PredicateOperator
 import me.ahoo.wow.api.query.QueryErrorCode
-import me.ahoo.wow.api.query.QueryException
 import me.ahoo.wow.api.query.QueryExpression
 import me.ahoo.wow.api.query.QueryPage
 import me.ahoo.wow.api.query.QuerySortDirection
@@ -40,6 +39,7 @@ import me.ahoo.wow.mongo.AggregateSchemaInitializer.toSnapshotCollectionName
 import me.ahoo.wow.mongo.Documents
 import me.ahoo.wow.mongo.Documents.replacePrimaryKeyToAggregateId
 import me.ahoo.wow.mongo.query.snapshot.SnapshotConditionConverter
+import me.ahoo.wow.query.QueryException
 import me.ahoo.wow.query.backend.QueryBackend
 import me.ahoo.wow.query.backend.SecuredQuery
 import me.ahoo.wow.query.schema.QuerySchema
