+++ wow-query/src/test/kotlin/me/ahoo/wow/query/QueryExceptionTest.kt
@@ -1,0 +1,36 @@
+/*
+ * Copyright [2021-present] [ahoo wang <ahoowang@qq.com> (https://github.com/Ahoo-Wang)].
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package me.ahoo.wow.query
+
+import me.ahoo.test.asserts.assert
+import me.ahoo.wow.api.query.QueryErrorCode
+import me.ahoo.wow.api.query.QueryStage
+import me.ahoo.wow.exception.WowException
+import org.junit.jupiter.api.Test
+
+class QueryExceptionTest {
+    @Test
+    fun `should expose the standard Wow error contract`() {
+        val cause = IllegalStateException("backend failed")
+        val exception = QueryException(QueryErrorCode.BACKEND_FAILURE, QueryStage.BACKEND, cause)
+
+        exception.assert().isInstanceOf(WowException::class.java)
+        exception.errorCode.assert().isEqualTo("BACKEND_FAILURE")
+        exception.errorMsg.assert().isEqualTo("Query failed: BACKEND_FAILURE at BACKEND.")
+        exception.message.assert().isEqualTo(exception.errorMsg)
+        exception.code.assert().isEqualTo(QueryErrorCode.BACKEND_FAILURE)
+        exception.stage.assert().isEqualTo(QueryStage.BACKEND)
+        exception.cause.assert().isSameAs(cause)
+    }
+}
