+++ wow-query/src/test/kotlin/me/ahoo/wow/query/gateway/SnapshotQueryGatewayTest.kt
@@ -27,7 +27,6 @@ import me.ahoo.wow.api.query.PredicateOperator
 import me.ahoo.wow.api.query.Query
 import me.ahoo.wow.api.query.QueryBudget
 import me.ahoo.wow.api.query.QueryErrorCode
-import me.ahoo.wow.api.query.QueryException
 import me.ahoo.wow.api.query.QueryExpression
 import me.ahoo.wow.api.query.QueryPage
 import me.ahoo.wow.api.query.QueryProjection
@@ -37,6 +36,7 @@ import me.ahoo.wow.api.query.RelativeTimeOperator
 import me.ahoo.wow.modeling.annotation.aggregateMetadata
 import me.ahoo.wow.modeling.metadata.AggregateMetadata
 import me.ahoo.wow.modeling.metadata.StateAggregateMetadata
+import me.ahoo.wow.query.QueryException
 import me.ahoo.wow.query.backend.QueryBackend
 import me.ahoo.wow.query.backend.QueryRouter
 import me.ahoo.wow.query.backend.SecuredQuery
