+++ wow-query/src/test/kotlin/me/ahoo/wow/query/policy/CompositeQueryPolicyTest.kt
@@ -23,11 +23,11 @@ import me.ahoo.wow.api.query.Query
 import me.ahoo.wow.api.query.QueryBudget
 import me.ahoo.wow.api.query.QueryCapabilities
 import me.ahoo.wow.api.query.QueryErrorCode
-import me.ahoo.wow.api.query.QueryException
 import me.ahoo.wow.api.query.QueryScope
 import me.ahoo.wow.api.query.QueryStage
 import me.ahoo.wow.api.query.SearchExpression
 import me.ahoo.wow.modeling.annotation.aggregateMetadata
+import me.ahoo.wow.query.QueryException
 import me.ahoo.wow.query.schema.JacksonQuerySchemaProvider
 import me.ahoo.wow.serialization.JsonSerializer
 import me.ahoo.wow.tck.mock.MockCommandAggregate
